requirements.txt
streamlit
pandas
plotly
requests
google-generativeai
openai
huggingface-hub
sseclient
fastapi
uvicorn
pydantic
python-dotenv